Moved the abort button and execution progress bar to their own div above the sequences
